Reorganize move action files
